+++ src/unifiedig/__init__.py
@@ -5,8 +5,9 @@ from importlib.metadata import PackageNotFoundError, version
 from .explainer import Explainer
 from .explanation import Explanation
 from .jax import JaxModel
+from .tensorflow import TensorFlowModel
 
-__all__ = ["Explainer", "Explanation", "JaxModel"]
+__all__ = ["Explainer", "Explanation", "JaxModel", "TensorFlowModel"]
 
 try:
     __version__ = version("unifiedig")